Added a regex expression for file cleaning
Added a regex expression that I just wrote which will remove everything accept the proxy's address and port. 
I am no expert on regex and there is for sure a better way to accomplish this task however this works for me and thats all that matters.
I just wanted to propose this in hopes someone will find it useful.